Change velocity to double

+++ src/com/MRS/NeckbeardEngine/Enemies/Iris.java
@@ -4,7 +4,7 @@ import com.MRS.NeckbeardEngine.*;
 
 public class Iris extends Enemy {
      
-     public Iris (State state, int x, int y, int yVelocity, Projectile projectileType, PowerUpPickup heldPowerUp, long timeLine) {
+     public Iris (State state, int x, int y, double yVelocity, Projectile projectileType, PowerUpPickup heldPowerUp, long timeLine) {
           super(state, x, y, 0, yVelocity, projectileType, heldPowerUp, timeLine);
           health = 2;
      }
